Move TimerDisplay, Progress, Next Button components
Timer display moved into components folder.
Progress and Next Button components removed and instead directly
returned from the Test render.
